pixie type 1 damage received

+++ Digitakomo/Assets/Script/Enemies/Shared/EnemyBaseClass.cs
@@ -16,8 +16,8 @@ public abstract class EnemyBaseClass : MonoBehaviour
     // Stats
     [Header("EnemyBase Class")]
     [SerializeField]
-    protected int hp = 1;
-    protected int originalHp = 1;
+    protected float hp = 1;
+    protected float originalHp = 1;
     [SerializeField]
     protected int damage = 1;
     [SerializeField]
@@ -100,7 +100,7 @@ public abstract class EnemyBaseClass : MonoBehaviour
     // Called to modify the HP
     // Can be decrease or Increase
     // When HP reaches 0, gameObject is deactivated
-    protected virtual void ModifyHealth(int modifyAmt)
+    protected virtual void ModifyHealth(float modifyAmt)
     {
         hp += modifyAmt;
         if (hp < 1)
